Organize files + Rename class
